Update main.js
Added a Configuration Section:

A configuration section was introduced at the beginning of the script. Here, basic settings like breakevenRoas, AverageCvr, impressionThreshold, daysAgo, and feedLabels can be adjusted.
Customizable Product Type Names:

Variables for naming different product types (productTypes) were added within the configuration section. These names are customizable and used in the logic for determining the product type.
The names for LowImpressions and Poor Performer were swapped.
Custom Label Selection:

An option to select the Custom Label (selectedCustomLabel) was added to the configuration section. This allows for the selection of a Custom Label ranging from 0 to 4.
Function to Set Custom Label in Google Sheet:

A function named setCustomLabelInSheet was added. This function sets the selected Custom Label in cell H1 of a specified Google Sheet ("EYL-Izer"). The function uses the existing SPREADSHEET_URL variable to access the spreadshee